Protect Resources with CEL (#46)
Ensure labels and annotations that are expected to exist do.  We could
also extend this to make them immutable, but I suspect if we need to
perform a migration or some form of emergency patching this would be an
impediment more than a benfit at the operations level.  One for the
future though when things get more stable and mature.