Optimize API Calls (#57)
Any handler (regions, clusters) that needed access to either the
identity service or the region service would unconditionally create a
client, and that would involve TLS handshakes and token exchange.  For
things like clusters, they only need to update allocations on
create/update/delete, not read, and only need access to flavors and
images on create/update, thus there is an implicit penalty for APIs that
don't need those clients.  This makes API access to those client lazy,
and they will only create a client if necessary.  Secondly, much like
the region service, we expect things like regions, flavors and images to
be relatively static, so we can put a cache in front of these calls with
an hour timeout.