Fix Multiple Clusters (#75)
Good catch by staging!  We want to filter servers and security groups by
cluster, not outright reject them if the cluster is a mismatch due to
the API list interface returning all resources related to the
organization and not scoped to a specific cluster.  We should perhaps
allow the list APIs to choose a tag selector and push this down in the
next release.